asi1_agent/requirements.txt
uagents
uagents-core>=0.2.0
python-dotenv
requests
openai
gspread
google-auth
google-api-python-client
